gun_values removal
the gun_values script was almost redundant so I changed gun variables to be stored by the gun object itself. for this i also added a free shotgun model for the shotgun rather than reusing the revolver. Currently there is a visual glitch when turning with the shotgun. Has only been implemented to the testground, so right now shooting will not work in the other scenes, will fix next